convert portal stream format in conference
